moved enums and hooks to models
