+++ src/servers/ZoneServer2016/classes/simpleConstruction.ts
@@ -8,7 +8,7 @@
 
 import { BaseLightweightCharacter } from "./baselightweightcharacter";
 import { ZoneServer2016 } from "../zoneserver";
-import { Items } from "../enums";
+import { Items } from "../models/enums";
 function getDamageRange(definitionId: number): number {
   switch (definitionId) {
     case Items.METAL_WALL:
