+++ src/servers/ZoneServer2016/models/enums.ts
